+++ tests/libtest/lib1506.c
@@ -42,8 +42,8 @@ static CURLcode test_lib1506(const char *URL)
 
   /* Create fake DNS entries for serverX.example.com for all handles */
   for(i = 0; i < CURL_ARRAYSIZE(curl); i++) {
-    curl_msnprintf(dnsentry, sizeof(dnsentry), "server%d.example.com:%s:%s",
-                   (int)i + 1, port, address);
+    curl_msnprintf(dnsentry, sizeof(dnsentry), "server%zu.example.com:%s:%s",
+                   i + 1, port, address);
     curl_mprintf("%s\n", dnsentry);
     slist2 = curl_slist_append(slist, dnsentry);
     if(!slist2) {
@@ -67,8 +67,8 @@ static CURLcode test_lib1506(const char *URL)
     easy_init(curl[i]);
     /* specify target */
     curl_msnprintf(target_url, sizeof(target_url),
-                   "http://server%d.example.com:%s/path/1506%04i",
-                   (int)i + 1, port, (int)i + 1);
+                   "http://server%zu.example.com:%s/path/1506%04zu",
+                   i + 1, port, i + 1);
     target_url[sizeof(target_url) - 1] = '\0';
     easy_setopt(curl[i], CURLOPT_URL, target_url);
     /* go verbose */
