+++ tests/libtest/lib2032.c
@@ -49,7 +49,7 @@ static size_t callback(char *ptr, size_t size, size_t nmemb, void *data)
   if(CURLE_OK != code) {
     curl_mfprintf(stderr, "%s:%d curl_easy_getinfo() failed, "
                   "with code %d (%s)\n",
-                  __FILE__, __LINE__, (int)code, curl_easy_strerror(code));
+                  __FILE__, __LINE__, code, curl_easy_strerror(code));
     ntlmcb_res = TEST_ERR_MAJOR_BAD;
     return failure;
   }
