+++ tests/libtest/lib2301.c
@@ -50,8 +50,8 @@ static size_t t2301_write_cb(char *b, size_t size, size_t nitems, void *p)
     0x8a, 0x0
   };
   size_t incoming = nitems;
-  curl_mfprintf(stderr, "Called CURLOPT_WRITEFUNCTION with %d bytes: ",
-                (int)nitems);
+  curl_mfprintf(stderr, "Called CURLOPT_WRITEFUNCTION with %zu bytes: ",
+                nitems);
   for(i = 0; i < nitems; i++)
     curl_mfprintf(stderr, "%02x ", (unsigned char)buffer[i]);
   curl_mfprintf(stderr, "\n");
